Set some Android services to nagara_hal class
By adding new class, it is now possible to start some HAL services after
init by composer is done. In particular, it was found to be important for
Media condec services. This approach can be extended later for other
services that seem to be not initialized properly on droid start.